requirements.txt
Pillow>=2.0.0
coverage
html5lib==1.0b10
httplib2==0.7.6
nose==1.3.3
pyPdf2==1.26
reportlab>=3.0
six
Sphinx==1.4.8
sphinx-rtd-theme==0.1.9adafruit-ampy==1.0.3
aiohttp==2.3.5
alabaster==0.7.10
aniso8601==1.2.0
apiai==1.2.3
appdirs==1.4.3
asn1crypto==0.22.0
async-timeout==2.0.0
attrs==17.4.0
Automat==0.6.0
Babel==2.5.1
beautifulsoup4==4.6.0
bs4==0.0.1
cairocffi==0.8.0
CairoSVG==2.0.3
certifi==2018.1.18
certify==1.0rc6
cffi==1.10.0
chardet==3.0.4
click==4.1
colorama==0.3.9
constantly==15.1.0
coverage==4.4.1
cryptography==2.1.4
cssselect==1.0.1
cssselect2==0.2.0
cx-Freeze==5.0.2
cycler==0.10.0
DateTime==4.2
docopt==0.6.2
docutils==0.14
dropbox==8.4.0
duckduckpy==0.2
ecdsa==0.13
emoji==0.4.5
esptool==2.3.1
et-xmlfile==1.0.1
fbmq==2.3.1
feedfinder2==0.0.4
feedparser==5.2.1
Flask==0.12.1
Flask-Ask==0.9.7
Flipkart-AffiliatesAPI==0.0.1.1
forismatic==1.0
future==0.16.0
gi==1.2
google==1.9.3
google-api-python-client==1.6.2
Google-Search-API==1.1.14
googletrans==2.1.4
html5lib==1.0b10
httplib2==0.10.3
hyperlink==18.0.0
idna==2.5
imagesize==0.7.1
imutils==0.4.5
incremental==17.5.0
inflect==0.2.5
itsdangerous==0.24
jaraco.itertools==2.0.1
jdcal==1.3
jieba3k==0.35.1
Jinja2==2.9.6
lib==3.0.0
lxml==4.1.1
MarkupSafe==1.0
matplotlib==2.1.0
messengerbot==0.1.4
more-itertools==3.2.0
multidict==3.1.3
mutagen==1.39
ndg-httpsclient==0.4.3
newspaper3k==0.2.2
nltk==3.2.4
nose==1.3.3
numpy==1.13.3
oauth2client==4.1.2
olefile==0.44
opencv-python==3.4.0.12
openpyxl==2.4.8
paho-mqtt==1.3.1
pandas==0.22.0
parsel==1.4.0
pdfkit==0.6.1
pdfminer3k==1.3.1
Pillow==4.2.1
pluggy==0.6.0
ply==3.11
pprint==0.1
prettyprint==0.1.5
prettytable==0.7.2
progressbar2==3.34.2
protobuf==3.4.0
py==1.5.2
py-bing-search==0.2.6
py-ms-cognitive==0.4.0
py2exe==0.9.2.2
pyaes==1.6.1
pyasn1==0.3.2
pyasn1-modules==0.0.11
pycparser==2.18
PyDispatcher==2.0.5
PyDrive==1.3.1
pygal==2.4.0
pygame==1.9.3
Pygments==2.2.0
PyInstaller==3.2.1
pyjokes==0.5.0
pymessenger==0.0.7.0
pyOpenSSL==17.5.0
pyparsing==2.2.0
pyPdf==1.13
PyPDF2==1.26.0
Pyphen==0.9.4
pypiwin32==220
pyserial==3.4
pysnap==0.1.1
pytest==3.4.1
python-dateutil==2.6.1
python-telegram-bot==9.0.0
python-utils==2.2.0
python-vlc==3.0.102
pytz==2017.2
PyYAML==3.12
queuelib==1.4.2
reportlab==3.4.0
requests==2.18.4
requests-file==1.4.2
requests-toolbelt==0.8.0
rsa==3.4.2
schedule==0.4.3
selenium==2.53.6
SimplePool==0.1
six==1.10.0
snap==0.5
snapchat-bots==0.1
snowballstemmer==1.2.1
Sphinx==1.4.8
sphinx-rtd-theme==0.1.9
sphinxcontrib-websupport==1.0.1
tabula-py==1.0.0
tinycss==0.4
tinycss2==0.6.0
tldextract==2.1.0
tmdbsimple==1.7.0
uber-rides==0.6.0
Unidecode==0.4.21
uritemplate==3.0.0
urllib3==1.22
vcrpy==1.11.1
virtualenv==15.1.0
w3lib==1.19.0
WeasyPrint==0.40
webencodings==0.5.1
Werkzeug==0.12.2
wget==3.2
wikipedia==1.4.0
wikiquote==0.1.5
wolframalpha==3.0
wordnik==2.1.3
wrapt==1.10.11
xhtml2pdf==0.2b1
xmltodict==0.11.0
yarl==0.12.0
zope.interface==4.4.2
